--- Gruntfile.js
@@ -1,22 +1,0 @@
-'use strict';
-
-module.exports = function(grunt) {
-  grunt.initConfig({
-    pkg: grunt.file.readJSON('package.json'),
-    babel: {
-      options: {
-        comments: false,
-        presets: ['minify']
-      },
-      dist: {
-        files: {
-          'n2words.min.js': 'n2words.js'
-        }
-      }
-    }
-  });
-
-  grunt.loadNpmTasks('grunt-babel');
-
-  grunt.registerTask('default', ['babel']);
-}
